Convert default (1900s) eviction date to 'nil'
Similarly to court date and nosp dates, we get back a default-looking
date (1st Jan 1900) if the field is missing in UH. This may be the
fault of the tenancy API, or it may be the fault of MSSQL Server -
either way, it's not desirable behaviour for us (as it confuses users
and makes our classification logic harder).

In the future, it would be good to have this filtering in a single
component rather than spread across multiple fields, but at the time of
writing, this is expected to fix some significant issues with
filtering, so I am happy to miss that cleanup in this commit.