Refactor (and fix!) nosp/sms classification logic
This commit attempts to make it clearer which condtions 'escape' out
a particular classification, e.g. "if it's got an agreement on, don't
sedd a nosp'.

As expected, this exposed some weird conditions/edge cases. This
commit includes fixing the behaviour for these cases, including one
for letter two.

Notably, there is still duplication across each classification, and
the tests are hard to read. I hope to address these in future commits.