Express positive 'court date' examples first
Unfortunately, the diff makes the intent of this commit difficult to
see, as these examples have a lot in common.

This commit pulls the *entire* two examples that result in an
`apply_for_court_date` classification to the top of this file, as I've
found starting with a positive example, then the exceptions, to be a
clearer way of explaining through examples.