DRY up court date examples, add description
This commit attempts to make these specs more readable without losing
the explanatory power of examples. It also does a little work on the
output of those specs, attempting to express a high-level description
of each case.

There may be a way to express these descriptions with more standard
rspec constructs, but this is a small change to test a theory - that
it's clearer to explain the behaviour through smaller examples.