Wait for DB to come up in testing/development
With a seperate MySQL container, the first time it boots it can take a
little while to populate the docker volume with the necessary data.

This commit adds dockerize, a helper script that waits for services to
be connectable before allowing the service to start.

Also simplified the startup commands for the service.

See:
* https://github.com/jwilder/dockerize
* https://docs.docker.com/compose/rails/