Initial stage of classification engine refactor
In this commit, as a team, we isolated the rules to determine what
action to return.

Previously the rules worked in a cascading format, each action in the
legal process was evaluated, from last to beginning. The first action
whose criteria were matched would be returned.

While this pattern was sufficient for the initial roll out of this
functionality, it will quickly prove difficult to extend this to cater
for agreements.

In true refactoring spirit we tried not to touch the tests during the
refactor, using them as a baseline to ensure that our refactor did not
change the outcome of the function. However, this was not possible in
one test where it seems the fixture was not quite set up correctly.