Check Data action when court outcome but no date
There's a case that v1 flags should send a NOSP however it is
covered under an existing suspended possession order.

This is another case of poor data within UH.

Updated a few other test cases where the scenarios describe a
particular action but are edge cases that shouldn't trigger that
action, instead a check_data should be returned. For each of these the
v1 portion has to be disabled