Add agreement amount from aragdet
Wire up SQL to join on aragdet to get amount of the agreement.

aragdet appears to contain an update to the agreement stored in UH. For
our purposes, we will take each of these as an individual agreement. MAA
agreements are immutable and should be 'Cancel and replace'd if changes
are required to be made