Cancel old aragdat agreements
When an agreement is updated in UH it creates a new aragdat row. We
only want to ensure that only the final aragdat row can have the status
as per the agreement. All previous ones should be cancelled, as they're
replaced by each new row.

Excuse the rather complicated looking code here. First in order to only
cancel previous changes of the same arag_ref we reduce the results into
a hash. Then we process each distinct arag, cancelling all changes with
exception to the final change which gets the original arag_status.