Use realisitc nil values as stored in UH
The UH 'NULL' for a date is 1900-01-01T00:00:00, of course. And likewise
the NULL for the court outcome is '   '.

Ideally this would be taken care of within the
universal_housing_criteria, but that would have implications on the
existing parts of the system which we don't really have time to sort at
the moment