Create new create_court_case_and_sync usecase to fix recursion issue
By creating this new usecase and calling it from the controller we avoid the sync from recursively
trying to sync the case causing the test suite to constantly run. Furthermore this commit now passes the
username independently from the court case request params for clarity as the username is not vital to creating
a court case.